Backend/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
dotenv
tiktoken
openai
azure-cosmos
azure-search-documents
azure-core
azure-identity
azure-ai-formrecognizer
azure-cognitiveservices-speech